resturcturing GET admin rest
