+++ controllers/api/adminController.js
@@ -1,0 +1,11 @@
+const db = require('../../models');
+const Restaurant = db.Restaurant;
+const Category = db.Category;
+const adminController = {
+  getRestaurants: (req, res) => {
+    return Restaurant.findAll({ include: [Category] }).then(restaurants => {
+      return res.json({ restaurants });
+    });
+  }
+};
+module.exports = adminController;
